+++ spec/spec_helper.rb
@@ -8,7 +8,10 @@ if ENV.fetch('COVERAGE', 'f').start_with? 't'
 end
 
 require 'rubocop-ast'
-RuboCop::AST::Builder.modernize if ENV['MODERNIZE']
+if ENV['MODERNIZE']
+  RuboCop::AST::Builder.modernize
+  RuboCop::AST::Builder.emit_forward_arg = false # inverse of default
+end
 
 RSpec.shared_context 'ruby 2.3', :ruby23 do
   let(:ruby_version) { 2.3 }